reviews_django/requirements.txt
Django==1.9.7
django-cors-headers==1.1.0
django-guardian==1.4.4
djangorestframework==3.3.3
djangorestframework-jsonapi==2.0.1
inflection==0.3.1
oauthlib==1.0.3
Pillow==3.2.0
requests==2.10.0
requests-oauth2==0.2.0
six==1.10.0
